Update 14:14 02/12/2022

+++ BaiLT_SQL_21520455_PhanTuanThanh/BaiLT_SQL_21520455_PhanTuanThanh/Form1.cs
@@ -7,6 +7,7 @@ using System.Drawing;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Web;
 using System.Windows.Forms;
 
 namespace BaiLT_SQL_21520455_PhanTuanThanh
@@ -16,6 +17,7 @@ namespace BaiLT_SQL_21520455_PhanTuanThanh
         public Form1()
         {
             InitializeComponent();
+
         }
 
         SqlConnection connection;
@@ -41,21 +43,54 @@ namespace BaiLT_SQL_21520455_PhanTuanThanh
             loaddata();
         }
 
-        private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
+        private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             DataGridView dgv = dataGridView1;
-            int i = dgv.CurrentRow.Index;
-            textBoxMSNV.Text = dgv.Rows[i].Cells[0].Value.ToString();
-            textBoxHoTen.Text = dgv.Rows[i].Cells[1].Value.ToString();
-            textBoxNgSinh.Text = dgv.Rows[i].Cells[2].Value.ToString();
-            textBoxNgVL.Text = dgv.Rows[i].Cells[3].Value.ToString();
+            int i = e.RowIndex;
+            if (i >= 0)
+            {
+                textBoxMSNV.Text = dgv.Rows[i].Cells[0].Value.ToString();
+                textBoxHoTen.Text = dgv.Rows[i].Cells[1].Value.ToString();
+                dateTimePickerNgSinh.Text = dgv.Rows[i].Cells[2].Value.ToString();
+                dateTimePickerNgVL.Text = dgv.Rows[i].Cells[3].Value.ToString();
+            }
         }
 
         private void buttonDelete_Click(object sender, EventArgs e)
         {
-            command.Connection.CreateCommand();
-            command.CommandText = "DELETE FROM NhanVien WHERE MSNV = '" + textBoxMSNV.Text + "'";
-            command.ExecuteNonQuery();
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc chắn muốn xóa không?", "Cảnh báo",
+                                        MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (dialogResult == DialogResult.Yes)
+            {
+                command.Connection.CreateCommand();
+                command.CommandText = "DELETE FROM NhanVien WHERE MSNV = '" + textBoxMSNV.Text + "'";
+                command.ExecuteNonQuery();
+                loaddata();
+            }
+        }
+
+        private void buttonUpdate_Click(object sender, EventArgs e)
+        {
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc chắn muốn sửa không?", "Cảnh báo",
+                                        MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (dialogResult == DialogResult.Yes)
+            {
+                string msnv = textBoxMSNV.Text;
+                string Ten = textBoxHoTen.Text;
+                DateTime NgSinh = DateTime.Parse(dateTimePickerNgSinh.Text);
+                DateTime NgVL = DateTime.Parse(dateTimePickerNgVL.Text);
+
+                command.Connection.CreateCommand();
+                command.CommandText = "UPDATE NhanVien SET HoTen = N'"+Ten+"', NgSinh = '"+NgSinh+"', NgVL = '"+NgVL+"' WHERE MSNV = '" + msnv + "'";
+                command.ExecuteNonQuery();
+                loaddata();
+            }   
+        }
+
+        private void buttonAdd_Click(object sender, EventArgs e)
+        {
+            FormAdd form = new FormAdd();
+            form.ShowDialog();
             loaddata();
         }
     }
